Add a base help message when no argument is given
When executing meliodas.py with no arguments, the interface would
fail due to an AttributeError because it was passing an empty
Namespace object to the functions.

In order to solve that, the 'sys' module was used to check the
arguments that were given and print a help message if no argument
is specified.